--- code/hyperdb/src/helpers/db/main.ts
@@ -1,1 +1,0 @@
-export * from './Db';
